requirements.txt
fastapi
couchdb
python-jose[cryptography]
passlib